==> 202107_EXP2/tag-seq/scripts/5_GOterm_Annotation_unique_multistressor_triploid.Rmd <==
---
title: "5_GOterm_Annotation"
author: "Matt George"
date: "2022-09-08"
output: html_document
---

```{r setup, include=TRUE}
knitr::opts_chunk$set(echo = TRUE, warning = TRUE, message = FALSE, cache = TRUE)
```

```{bash, engine.path="C:/Windows/System32/bash.exe"}
echo $PATH
```
# Make file structure
```{bash engine.path="C:/Windows/System32/bash.exe", include=FALSE}

mkdir GOterm_annotation/c-ontrol_ploidy/
mkdir GOterm_annotation/h-eat_ploidy/
mkdir GOterm_annotation/d-esiccation_ploidy/
mkdir GOterm_annotation/c-ontrol_ploidy/
mkdir GOterm_annotation/d-iploid_desiccation/
mkdir GOterm_annotation/t-riploid_heat/
mkdir GOterm_annotation/t-riploid_desiccation/

```

### Load libraries
```{r load_libraries, inlcude = TRUE}

## clear
rm(list=ls())

## Install Rtools directly from (https://cran.r-project.org/bin/windows/Rtools/), then install these on first run:
# BiocManager::install(version = "3.13")
# BiocManager::install("BiocGenerics")
# BiocManager::install("graph")
# BiocManager::install("Biobase")
# BiocManager::install("AnnotationDbi")
# BiocManager::install("GO.db")
# BiocManager::install("topGO", force = TRUE)
# BiocManager::install("simplifyEnrichment") #Needs newest R Version (4.1.0)
# BiocManager::install("rrvgo", force = TRUE)

# List of packages we want to install (run every time)
load.lib<-c("tidyverse","RColorBrewer","readxl","ggpubr","beepr","BiocManager","topGO","simplifyEnrichment","SparseM","magick","gclus","rrvgo")

# Select only the packages that aren't currently installed (run every time)
install.lib <- load.lib[!load.lib %in% installed.packages()]

# And finally we install the missing packages, including their dependency.
for(lib in install.lib) install.packages(lib,dependencies=TRUE)
# After the installation process completes, we load all packages.
sapply(load.lib,require,character=TRUE)
                        
```
#Set ggplot theme
```{r ggplot_theme, include=FALSE}

my_theme <- theme(line              = element_line(size=1),
                  rect              = element_rect(size=1),
                  text              = element_text(size=9,color="black"),
                  panel.background  = element_blank(),
                  panel.grid.major  = element_blank(), 
                  panel.grid.minor  = element_blank(),
                  axis.text.x       = element_text(size=9,color="black"),
                  axis.text.y       = element_text(size=9,color="black"),
                  axis.title.x      = element_text(margin = margin(t = 10, r = 0, b = 0, l = 0)),
                  axis.title.y      = element_text(margin = margin(t = 0, r = 10, b = 0, l = 0)),
                  axis.ticks.x      = element_line(color="black"),
                  axis.ticks.y      = element_line(color="black"),
                  # axis.line         = element_line(color = "black", size = 0.1),
                  panel.border      = element_rect(color = "black", fill=NA, size=1),
                  legend.key        = element_blank()) # removes background of legend bullets
```

### NOTES ####################################################################################################

# Chandra Rajan et al. (2020) Methods:
# https://onlinelibrary.wiley.com/action/downloadSupplement?doi=10.1111%2Fgcb.15675&file=gcb15675-sup-0002-FileS2.txt

# Bioconductor SimplifyEnrichment Methods:
# http://www.bioconductor.org/packages/devel/bioc/vignettes/simplifyEnrichment/inst/doc/simplifyEnrichment.html
# https://www.biostars.org/p/350710/#350712
# https://bioconductor.org/packages/devel/bioc/vignettes/topGO/inst/doc/topGO.pdf
# https://www.bioconductor.org/packages/devel/bioc/vignettes/rrvgo/inst/doc/rrvgo.html

# Yaamini's Notebook
# https://github.com/RobertsLab/project-oyster-oa/blob/master/code/Haws/08-GOterm-Annotation.ipynb

### ####################################################################################################

# Import list of DEG lists and associated GO term annotations
```{r}
triploid_MS_unique <- read.table("gene_tables/filtered/HISAT2_multiqc_biplot/triploid_MS_unique.txt", 
                               header = TRUE, sep = "\t") 
nrow(triploid_MS_unique) # 10963

triploid_MS_unique_bg <- read.delim("gene_tables/filtered/HISAT2_multiqc_biplot/triploid_desiccation_bg.txt", 
                               header = TRUE, sep = "\t") 
nrow(triploid_MS_unique_bg) # 577,330

```

```{r}
# Filter by BP, CC, or MF
triploid_MS_unique_BP <- triploid_MS_unique %>% filter(., triploid_MS_unique$GOcat == "P")
triploid_MS_unique_CC <- triploid_MS_unique %>% filter(., triploid_MS_unique$GOcat == "C")
triploid_MS_unique_MF <- triploid_MS_unique %>% filter(., triploid_MS_unique$GOcat == "F")
nrow(triploid_MS_unique_BP) # 4806

# Count GOIDs
triploid_MS_unique_BP_count <- as.data.frame(table(triploid_MS_unique_BP$GOID))
triploid_MS_unique_CC_count <- as.data.frame(table(triploid_MS_unique_CC$GOID))
triploid_MS_unique_MF_count <- as.data.frame(table(triploid_MS_unique_MF$GOID))

nrow(unique(triploid_MS_unique_BP_count)) # 910 of unique GOIDs associated with ploidy
nrow(unique(triploid_MS_unique_CC_count)) # 237 of unique GOIDs associated with ploidy
nrow(unique(triploid_MS_unique_MF_count)) # 366 of unique GOIDs associated with ploidy

# Add column names
colnames(triploid_MS_unique_BP_count) <- c("GOID", "Freq")
colnames(triploid_MS_unique_CC_count) <- c("GOID", "Freq")
colnames(triploid_MS_unique_MF_count) <- c("GOID", "Freq")

# Filter by frequency
triploid_MS_unique_BP_count <- triploid_MS_unique_BP_count %>% filter(triploid_MS_unique_BP_count$Freq != 0)
triploid_MS_unique_CC_count <- triploid_MS_unique_CC_count %>% filter(triploid_MS_unique_CC_count$Freq != 0)
triploid_MS_unique_MF_count <- triploid_MS_unique_MF_count %>% filter(triploid_MS_unique_MF_count$Freq != 0)

# Save GOterms associated with BP, CC, or MF
write.csv(triploid_MS_unique_BP, "GOterm_annotation/triploid_MS_unique/triploid_MS_unique-BP-ALL_GOterms.csv", quote = FALSE, row.names = FALSE)
write.csv(triploid_MS_unique_CC, "GOterm_annotation/triploid_MS_unique/triploid_MS_unique-CC-ALL_GOterms.csv", quote = FALSE, row.names = FALSE)
write.csv(triploid_MS_unique_MF, "GOterm_annotation/triploid_MS_unique/triploid_MS_unique-MF-ALL_GOterms.csv", quote = FALSE, row.names = FALSE)

triploid_MS_unique_BP_revigo <- triploid_MS_unique_BP[,c(10,2)]
triploid_MS_unique_BP_revigo <- unique(triploid_MS_unique_BP_revigo)

write.table(triploid_MS_unique_BP_revigo, "GOterm_annotation/triploid_MS_unique/triploid_MS_unique_BP_revigo.csv", col.names = FALSE, row.names = FALSE, quote = FALSE,  sep = "\t")

```

# Generate backgroud gene2go list
```{bash engine.path="C:/Windows/System32/bash.exe"}
pwd
#Extract transcript IDs
#Filter and save unique IDs
cut -f5 gene_tables/filtered/HISAT2_multiqc_biplot/triploid_desiccation_bg.txt | uniq \
> gene_tables/filtered/HISAT2_multiqc_biplot/triploid_MS_unique_bg_transcriptIDs_unique.txt

```

```{r}
triploid_MS_unique_bg_unique <- read.delim("gene_tables/filtered/HISAT2_multiqc_biplot/triploid_MS_unique_bg_transcriptIDs_unique.txt", header = TRUE, sep = "\t")
colnames(triploid_MS_unique_bg_unique) <- c("GeneID")
triploid_MS_unique_bg_unique$GeneID <- as.character(triploid_MS_unique_bg_unique$GeneID)

geneid2go <- read.delim("blast/geneid2go.tab", header = FALSE, sep = "\t")
colnames(geneid2go) <- c("GeneID","GOterm")
geneid2go <- unique(geneid2go)

nrow(triploid_MS_unique_bg_unique)

triploid_MS_unique_bg_unique <- left_join(triploid_MS_unique_bg_unique, geneid2go, by = "GeneID")

write.table(triploid_MS_unique_bg_unique, "GOterm_annotation/triploid_MS_unique/triploid_MS_unique_bg_unique.txt", sep = "\t", quote = FALSE, row.names = FALSE, col.names = FALSE)

nrow(triploid_MS_unique_bg_unique)

```

```{r}
########### Load genetic background #################
# Load data for gene enrichment
# Loading the GO annotations and GeneIDs. Each line has one transcript ID and all associated GOterms
triploid_MS_unique_bg_topGO <- topGO::readMappings(file = "GOterm_annotation/triploid_MS_unique/triploid_MS_unique_bg_unique.txt")
length(triploid_MS_unique_bg_topGO) #34907

str(head(triploid_MS_unique_bg_topGO)) #Confirm file structure
geneNames <- names(triploid_MS_unique_bg_topGO) #Extract names to use as gene universe
head(triploid_MS_unique_bg_topGO)
```

```{r}
triploid_MS_unique_genes     <- unique(triploid_MS_unique$product_accession) #Extract transcript ID
triploid_MS_unique_gene_list <- factor(as.integer(geneNames %in% triploid_MS_unique_genes))
names(triploid_MS_unique_gene_list) <- geneNames #PLOIDYGenes
str(triploid_MS_unique_gene_list)

```

########## Run TopGo #########################
```{r}
# Create topGOdata - Biological processes
triploid_MS_unique_topGO <- new("topGOdata", ontology = "BP", allGenes = triploid_MS_unique_gene_list,
                            annot = annFUN.gene2GO, gene2GO = triploid_MS_unique_bg_topGO) #Create biological process topGO object
triploid_MS_unique_topGO #Get summary of object

```

#Create a statistical results table with statistical test results. Order by p-value (classic), and include all results (topNodes)
```{r}
test.stat <- new("classicCount", testStatistic = GOFisherTest, name = "Fisher test")

# PLOIDY
resultFisher.triploid_MS_unique_BP <- getSigGroups(triploid_MS_unique_topGO, test.stat)
pvalFis.triploid_MS_unique_BP <- score(resultFisher.triploid_MS_unique_BP) #Extract p-values
hist(pvalFis.triploid_MS_unique_BP, 50, xlab = "p-values") #Plot histogram of p-values

allRes.triploid_MS_unique_BP <- GenTable(triploid_MS_unique_topGO, 
                                     classic = resultFisher.triploid_MS_unique_BP, 
                                     ranksOf = "classic", 
                                     orderBy = "classic", 
                                     topNodes = length(pvalFis.triploid_MS_unique_BP)) 
head(allRes.triploid_MS_unique_BP)

write.csv(allRes.triploid_MS_unique_BP, "GOterm_annotation/triploid_MS_unique/triploid_MS_unique-BP-FisherTestResults.csv", quote = FALSE, row.names = FALSE) #Save dataframe

```

```{r}
############# Match enriched GOterms with general annotation information and product ID #################
# Filter significantly enriched GOterms
sigRes.triploid_MS_unique_BP <- filter(allRes.triploid_MS_unique_BP,classic<0.01)
write.csv(sigRes.triploid_MS_unique_BP, "GOterm_annotation/triploid_MS_unique/sigRes-Enriched-GOterms-triploid_MS_unique_BP.csv", quote = FALSE, row.names = FALSE) #Save dataframe

# keep GOID, GOterm, and p-value
sigRes.triploid_MS_unique_BP <- sigRes.triploid_MS_unique_BP[,c(1,6)] #Filter significantly enriched GOterms, only keep GOID and p-value
colnames(sigRes.triploid_MS_unique_BP) <- c("GOID", "p.value") #Change column names

# Add annotation information
sigRes.triploid_MS_unique_BP_Annot <- merge(sigRes.triploid_MS_unique_BP, triploid_MS_unique_BP, by = "GOID") #Additional annotations
length(unique(sigRes.triploid_MS_unique_BP_Annot$gene)) #50 unique genes
length(unique(sigRes.triploid_MS_unique_BP_Annot$product_accession)) #85 unique transcripts
head(sigRes.triploid_MS_unique_BP_Annot) #Confirm formatting

write.csv(sigRes.triploid_MS_unique_BP_Annot, "GOterm_annotation/triploid_MS_unique/sigRes-Enrichched-triploid_MS_unique_BP-withTranscript.csv", quote = FALSE, row.names = FALSE) #Save dataframe

sigRes.triploid_MS_unique_BP_Annot_noT <- unique(sigRes.triploid_MS_unique_BP_Annot[,-7]) #Drop transcript column and keep only unique rows
head(sigRes.triploid_MS_unique_BP_Annot_noT) #Confirm formatting
write.csv(sigRes.triploid_MS_unique_BP_Annot_noT, "GOterm_annotation/triploid_MS_unique/sigRes-Enrichched-triploid_MS_unique_BP-withoutTranscript.csv", quote = FALSE, row.names = FALSE) #Save dataframe

```

```{r}
## Run topGO + KS test
all_genes       <- unique(triploid_MS_unique[,c(5,2)]) #Extract transcript ID
DEG_list        <- all_genes$padj
names(DEG_list) <- all_genes$product_accession
# DEG_list        <- factor(as.integer(geneNames %in% DEG_list))
# all_genes_list <- factor(as.integer(geneNames %in% all_genes))
# names(all_genes_list) <- geneNames #PLOIDYGenes
# str(all_genes_list)

selection <- function(allScore){ return(allScore < 0.05)} # function that returns TRUE/FALSE for p-values<0.05

GOdata <- new("topGOdata",
              ontology = "BP",
              allGenes = DEG_list,
              annot    = annFUN.gene2GO,
              gene2GO  = triploid_MS_unique_bg_topGO,
              geneSel  = selection,
              nodeSize = 10)

results.ks <- runTest(GOdata, algorithm="classic", statistic="ks")

```

```{r}

# Generate summary table of ks scores
goEnrichment      <- GenTable(GOdata, KS=results.ks, orderBy="KS", topNodes = length(results.ks@score))
for (i in 1:length(goEnrichment$KS)){if(goEnrichment$KS[i] == "< 1e-30"){goEnrichment$KS[i] <- 1e-10}}
goEnrichment$KS   <- as.numeric(goEnrichment$KS) # not sure why goEnrichment$KS is text - seems dumb
goEnrichment      <- goEnrichment[goEnrichment$KS<0.05,] # Filter by significant
goEnrichment      <- goEnrichment[,c("GO.ID","Term","KS")] # Change column titles

goEnrichment$Term <- gsub("[a-z]*\\.\\.\\.$", "", goEnrichment$Term) # [a-z]*
# goEnrichment$Term <- gsub("\\.\\.\\.\\.\\.\\.\\.$", "", goEnrichment$Term)
goEnrichment$Term <- paste(goEnrichment$GO.ID, goEnrichment$Term, sep=", ")
goEnrichment$Term <- factor(goEnrichment$Term, levels=rev(goEnrichment$Term))
# goEnrichment      <- goEnrichment[1:20,] # Plot only top 20
goEnrichment$cat  <- 'other'# cat = GOterm keyword category

# Assign color labels by GOterm keyword
# add cat column w/ GOterm keyword
regulation      <- with(goEnrichment, subset(goEnrichment, subset = grepl(glob2rx("*regulation*"), Term)))
transcription   <- with(goEnrichment, subset(goEnrichment, subset = grepl(glob2rx("*transcription*"), Term)))
translation     <- with(goEnrichment, subset(goEnrichment, subset = grepl(glob2rx("*translation*"), Term)))
development     <- with(goEnrichment, subset(goEnrichment, subset = grepl(glob2rx("*develop*"), Term)))
morphogenesis   <- with(goEnrichment, subset(goEnrichment, subset = grepl(glob2rx("*morpho*"), Term)))
differentiation <- with(goEnrichment, subset(goEnrichment, subset = grepl(glob2rx("*diff*"), Term)))
metabolism      <- with(goEnrichment, subset(goEnrichment, subset = grepl(glob2rx("*metabol*"), Term)))
immune          <- with(goEnrichment, subset(goEnrichment, subset = grepl(glob2rx("*immune*"), Term)))


if(nrow(regulation)      > 0){regulation$cat       <- "regulation"}
if(nrow(transcription)   > 0){transcription$cat    <- "regulation"}
if(nrow(translation)     > 0){translation$cat      <- "regulation"}
if(nrow(development)     > 0){development$cat      <- "development"}
if(nrow(differentiation) > 0){differentiation$cat  <- "development"}
if(nrow(morphogenesis)   > 0){morphogenesis$cat    <- "development"}
if(nrow(metabolism)      > 0){metabolism$cat       <- "metabolism"}
if(nrow(immune)          > 0){immune$cat           <- "immune"}

# Find
goEnrichment <- anti_join(goEnrichment, regulation,      by = "GO.ID") %>% bind_rows(regulation)
goEnrichment <- anti_join(goEnrichment, transcription,   by = "GO.ID") %>% bind_rows(transcription)
goEnrichment <- anti_join(goEnrichment, translation,     by = "GO.ID") %>% bind_rows(translation)
goEnrichment <- anti_join(goEnrichment, development,     by = "GO.ID") %>% bind_rows(development)
goEnrichment <- anti_join(goEnrichment, differentiation, by = "GO.ID") %>% bind_rows(differentiation)
goEnrichment <- anti_join(goEnrichment, morphogenesis,   by = "GO.ID") %>% bind_rows(morphogenesis)
goEnrichment <- anti_join(goEnrichment, metabolism,      by = "GO.ID") %>% bind_rows(metabolism)
goEnrichment <- anti_join(goEnrichment, immune,          by = "GO.ID") %>% bind_rows(immune)


# Manual assignment
# goEnrichment$cat[1:2] <- 'regulation'
goEnrichment <- goEnrichment %>% arrange((KS))

# Colors
# regulation = "royalblue1"
# metabolism = "palegreen1"
# development = "salmon1"
# other = "gray79"


# Plot
p1 <- ggplot(goEnrichment, aes(x=Term, y=-log10(KS), fill = cat)) +
             stat_summary(geom = "bar", fun = mean, position = "dodge", color="black", size = 0.3) +
             xlab("Biological process") +
             ylab("Enrichment") +
             # scale_fill_manual(values=c("salmon1","royalblue1")) +
             # scale_y_continuous(breaks = seq(0, 8, by = 2), limits = c(0,9)) +
             # guides(colour=guide_legend(override.aes=list(size=2.5))) +
             coord_flip() +
             my_theme

p1

bp1  <- ggplot(goEnrichment, aes(x=factor(1),y=cat, fill=cat)) +
              geom_bar(width = 1, stat = "identity", show.legend = TRUE) +
              # scale_fill_manual(values = c("lightskyblue","indianred")) + labs(NULL) +
              coord_polar("y", start=0) + my_theme
bp1

ggsave("GOterm_annotation/triploid_MS_unique/triploid_MS_unique_Kolmogorov-Smirnov-BP.png",
       plot   = p1,
       dpi    = 600,
       device = "png",
       width  = 7,
       height = 12,
       units  = "in")

ggsave("GOterm_annotation/triploid_MS_unique/triploid_MS_unique_pie-chart-BP.png",
       plot   = bp1,
       dpi    = 2500,
       device = "png",
       width  = 6,
       height = 5,
       units  = "in")


```

```{r}
## Run topGO + KS test
all_genes       <- unique(triploid_MS_unique[,c(5,2)]) #Extract transcript ID
DEG_list        <- all_genes$padj
names(DEG_list) <- all_genes$product_accession
# DEG_list        <- factor(as.integer(geneNames %in% DEG_list))
# all_genes_list <- factor(as.integer(geneNames %in% all_genes))
# names(all_genes_list) <- geneNames #PLOIDYGenes
# str(all_genes_list)

selection <- function(allScore){ return(allScore < 0.05)} # function that returns TRUE/FALSE for p-values<0.05

GOdata <- new("topGOdata",
              ontology = "CC",
              allGenes = DEG_list,
              annot    = annFUN.gene2GO,
              gene2GO  = triploid_MS_unique_bg_topGO,
              geneSel  = selection,
              nodeSize = 10)

results.ks <- runTest(GOdata, algorithm="classic", statistic="ks")

```

```{r}

# Generate summary table of ks scores
goEnrichment      <- GenTable(GOdata, KS=results.ks, orderBy="KS", topNodes = length(results.ks@score))
goEnrichment$KS   <- as.numeric(goEnrichment$KS) # not sure why goEnrichment$KS is text - seems dumb
goEnrichment      <- goEnrichment[goEnrichment$KS<0.05,] # Filter by significant
goEnrichment      <- goEnrichment[,c("GO.ID","Term","KS")] # Change column titles

goEnrichment$Term <- gsub("[a-z]*\\.\\.\\.$", "", goEnrichment$Term) # [a-z]*
# goEnrichment$Term <- gsub("\\.\\.\\.\\.\\.\\.\\.$", "", goEnrichment$Term)
goEnrichment$Term <- paste(goEnrichment$GO.ID, goEnrichment$Term, sep=", ")
goEnrichment$Term <- factor(goEnrichment$Term, levels=rev(goEnrichment$Term))
# goEnrichment      <- goEnrichment[1:20,] # Plot only top 20
goEnrichment$cat  <- 'other'# cat = GOterm keyword category

# Assign color labels by GOterm keyword
# add cat column w/ GOterm keyword
metabolism    <- with(goEnrichment, subset(goEnrichment, subset = grepl(glob2rx("*metabol*"), Term)))
if(nrow(metabolism) > 0){metabolism$cat <- "metabolism"} 
regulation      <- with(goEnrichment, subset(goEnrichment, subset = grepl(glob2rx("*regulation*"), Term)))
if(nrow(regulation) > 0){regulation$cat  <- "regulation"}
development     <- with(goEnrichment, subset(goEnrichment, subset = grepl(glob2rx("*develop*"), Term)))
if(nrow(development) > 0){development$cat <- "development"}
differentiation <- with(goEnrichment, subset(goEnrichment, subset = grepl(glob2rx("*diff*"), Term)))
if(nrow(differentiation) > 0){differentiation$cat <- "development"}

# Find 
goEnrichment <- anti_join(goEnrichment, metabolism, by = "GO.ID") %>% bind_rows(metabolism)
goEnrichment <- anti_join(goEnrichment, development, by = "GO.ID") %>% bind_rows(development)
goEnrichment <- anti_join(goEnrichment, differentiation, by = "GO.ID") %>% bind_rows(differentiation)
goEnrichment <- anti_join(goEnrichment, regulation, by = "GO.ID") %>% bind_rows(regulation)

# Manual assignment
# goEnrichment$cat[1:2] <- 'regulation'

# Colors
# regulation = "royalblue1"
# metabolism = "palegreen1"
# development = "salmon1"
# other = "gray79"

# Plot
p1 <- ggplot(goEnrichment, aes(x=Term, y=-log10(KS), fill = cat)) +
             stat_summary(geom = "bar", fun = mean, position = "dodge", color="black", size = 0.3) +
             xlab("Biological process") +
             ylab("Enrichment") +
             scale_fill_manual(values=c("salmon1","royalblue1")) +
             scale_y_continuous(breaks = seq(0, 8, by = 2), limits = c(0,9)) +
             # guides(colour=guide_legend(override.aes=list(size=2.5))) +
             coord_flip() +
             my_theme

p1

ggsave("GOterm_annotation/triploid_MS_unique/triploid_MS_unique_Kolmogorov-Smirnov-CC.png",
       plot   = p1,
       dpi    = 600,
       device = "png",
       width  = 7,
       height = 5,
       units  = "in")
```


```{r}
## Run topGO + KS test
all_genes       <- unique(triploid_MS_unique[,c(5,2)]) #Extract transcript ID
DEG_list        <- all_genes$padj
names(DEG_list) <- all_genes$product_accession
# DEG_list        <- factor(as.integer(geneNames %in% DEG_list))
# all_genes_list <- factor(as.integer(geneNames %in% all_genes))
# names(all_genes_list) <- geneNames #PLOIDYGenes
# str(all_genes_list)

selection <- function(allScore){ return(allScore < 0.05)} # function that returns TRUE/FALSE for p-values<0.05

GOdata <- new("topGOdata",
              ontology = "MF",
              allGenes = DEG_list,
              annot    = annFUN.gene2GO,
              gene2GO  = triploid_MS_unique_bg_topGO,
              geneSel  = selection,
              nodeSize = 10)

results.ks <- runTest(GOdata, algorithm="classic", statistic="ks")

```

```{r}

# Generate summary table of ks scores
goEnrichment      <- GenTable(GOdata, KS=results.ks, orderBy="KS", topNodes = length(results.ks@score))
goEnrichment$KS   <- as.numeric(goEnrichment$KS) # not sure why goEnrichment$KS is text - seems dumb
goEnrichment      <- goEnrichment[goEnrichment$KS<0.05,] # Filter by significant
goEnrichment      <- goEnrichment[,c("GO.ID","Term","KS")] # Change column titles

goEnrichment$Term <- gsub("[a-z]*\\.\\.\\.$", "", goEnrichment$Term) # [a-z]*
# goEnrichment$Term <- gsub("\\.\\.\\.\\.\\.\\.\\.$", "", goEnrichment$Term)
goEnrichment$Term <- paste(goEnrichment$GO.ID, goEnrichment$Term, sep=", ")
goEnrichment$Term <- factor(goEnrichment$Term, levels=rev(goEnrichment$Term))
# goEnrichment      <- goEnrichment[1:20,] # Plot only top 20
goEnrichment$cat  <- 'other'# cat = GOterm keyword category

# Assign color labels by GOterm keyword
# add cat column w/ GOterm keyword
metabolism    <- with(goEnrichment, subset(goEnrichment, subset = grepl(glob2rx("*metabol*"), Term)))
if(nrow(metabolism) > 0){metabolism$cat <- "metabolism"} 
regulation      <- with(goEnrichment, subset(goEnrichment, subset = grepl(glob2rx("*regulation*"), Term)))
if(nrow(regulation) > 0){regulation$cat  <- "regulation"}
development     <- with(goEnrichment, subset(goEnrichment, subset = grepl(glob2rx("*develop*"), Term)))
if(nrow(development) > 0){development$cat <- "development"}
differentiation <- with(goEnrichment, subset(goEnrichment, subset = grepl(glob2rx("*diff*"), Term)))
if(nrow(differentiation) > 0){differentiation$cat <- "development"}

# Find 
goEnrichment <- anti_join(goEnrichment, metabolism, by = "GO.ID") %>% bind_rows(metabolism)
goEnrichment <- anti_join(goEnrichment, development, by = "GO.ID") %>% bind_rows(development)
goEnrichment <- anti_join(goEnrichment, differentiation, by = "GO.ID") %>% bind_rows(differentiation)
goEnrichment <- anti_join(goEnrichment, regulation, by = "GO.ID") %>% bind_rows(regulation)

# Manual assignment
goEnrichment$cat[1:2] <- 'regulation'

# Colors
# regulation = "royalblue1"
# metabolism = "palegreen1"
# development = "salmon1"
# other = "gray79"

# Plot
p1 <- ggplot(goEnrichment, aes(x=Term, y=-log10(KS), fill = cat)) +
             stat_summary(geom = "bar", fun = mean, position = "dodge", color="black", size = 0.3) +
             xlab("Biological process") +
             ylab("Enrichment") +
             scale_fill_manual(values=c("salmon1","royalblue1")) +
             scale_y_continuous(breaks = seq(0, 8, by = 2), limits = c(0,9)) +
             # guides(colour=guide_legend(override.aes=list(size=2.5))) +
             coord_flip() +
             my_theme

p1

ggsave("GOterm_annotation/triploid_MS_unique/triploid_MS_unique_Kolmogorov-Smirnov-MF.png",
       plot   = p1,
       dpi    = 600,
       device = "png",
       width  = 7,
       height = 5,
       units  = "in")
```

## Semantic Similarity Plot
```{r eval=FALSE, include=FALSE}
# Calculate the semantic similarity matrix using the Rel method (default)
mat_triploid_MS_unique_BP <- GO_similarity(go_id = sigRes.triploid_MS_unique_BP$GOID, ont = "BP")

scores <- setNames(-log10(as.numeric(sigRes.triploid_MS_unique_BP$p.value)), sigRes.triploid_MS_unique_BP$GOID)
reducedTerms <- reduceSimMatrix(mat_triploid_MS_unique_BP,
                                scores,
                                threshold=0.7,
                                orgdb=triploid_MS_unique_bg_topGO)
heatmapPlot(mat_triploid_MS_unique_BP,
            reducedTerms,
            annotateParent=TRUE,
            annotationLabel="parentTerm",
            fontsize=6)

```

```{r eval=FALSE, include=FALSE}
pdf("GOterm_annotation/triploid_MS_unique/triploid_MS_unique-simplifyEnrichment-BP.pdf", width = 13, height = 9) #Save figure
simplifyGO(mat_triploid_MS_unique_BP,
           method = "binary_cut",
           # column_title = "", 
           # col = rev(plotColors), 
           fontsize_range = c(10,25),
           word_cloud_grob_param = list(max_width = 100)) #Plot GOterms based on semantic similarity. Do not include a column title. Set colors to be plot colors, and set fontsize to range from 10 to 40. Pass arguments to word_cloud_grob_param to dictate the colors of the words and maximum width
dev.off()
```

```{r eval=FALSE, include=FALSE}
df_triploid_MS_unique_BP <- simplifyGO(mat_triploid_MS_unique_BP, plot = FALSE) #Cluster GOterms by semantic similarity
head(df_triploid_MS_unique_BP) #Dataframe with GOIDs, GOterms, and cluster
sort(table(df_triploid_MS_unique_BP$cluster)) #Obtain size of clusters

```
